# Fix performance of images (and export)